moved errorboundary under shared folder
