+++ src/components/Layout.tsx
@@ -1,5 +1,5 @@
 import { Outlet, useNavigate } from "react-router-dom";
-import ErrorBoundary from "./ErrorBoundary";
+import ErrorBoundary from "./shared/ErrorBoundary";
 import Header from "./Header";
 import Footer from "./Footer";
 import { CSSProperties } from "react";
